refactor: reorganize codebase into feature modules
- Add feature modules for trips, activities, chat, auth, voting
- Move CreateTripModal, CreateActivityModal, TripChat into features
- Update pages to import from features and adjust i18n keys
- Mark Issue #3b (feature-based architecture) as completed in ISSUES.md
- Align Vitest config aliases with Vite config

Made-with: Cursor
