+++ src/features/activities/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Activities feature – public API.
+ * Import from '@/features/activities' only; do not reach into internal paths.
+ */
+
+export { default as CreateActivityModal } from './components/CreateActivityModal';
