+++ src/features/auth/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Auth feature – public API.
+ * Login/Signup forms and auth logic live here.
+ * Import from '@/features/auth' only; do not reach into internal paths.
+ *
+ * TODO: Extract LoginForm and SignupForm from pages when refactoring auth screens.
+ */
+
+// Placeholder: no exports yet. Pages still own auth UI; will be extracted in a follow-up.
