+++ src/features/chat/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Chat feature – public API.
+ * Import from '@/features/chat' only; do not reach into internal paths.
+ */
+
+export { default as TripChat } from './components/TripChat';
