+++ src/features/trips/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Trips feature – public API.
+ * Import from '@/features/trips' only; do not reach into internal paths.
+ */
+
+export { default as CreateTripModal } from './components/CreateTripModal';
