+++ src/features/voting/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Voting feature – public API.
+ * Voting UI and hooks will live here (Issue #10).
+ * Import from '@/features/voting' only; do not reach into internal paths.
+ *
+ * TODO: Add VoteButtons, scenario voting, decision view when implementing Issue #10.
+ */
+
+// Placeholder: no exports yet.
